122: Added if check to plantilles/nextcloud-restart.sh

+++ debian/linkat-servidor/usr/share/linkat/linkat-servidor/plantilles/nextcloud.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
 # Apparmor Nextcloud
-sudo ln -s /etc/apparmor.d/usr.lib.snapd.snap-confine.real /etc/apparmor.d/disable/
-sudo apparmor_parser -R /etc/apparmor.d/usr.lib.snapd.snap-confine.real
-sudo snap refresh core
+apparmorLink="/etc/apparmor.d/disable/usr.lib.snapd.snap-confine.real"
+if [[ ! -e $apparmorLink ]]; then
+        sudo ln -s /etc/apparmor.d/usr.lib.snapd.snap-confine.real /etc/apparmor.d/disable/
+        sudo apparmor_parser -R /etc/apparmor.d/usr.lib.snapd.snap-confine.real
+fi
 
 
 # Install and Configure Nextcloud via snap
